+++ tests/AbstractCompareTestCase.php
@@ -17,16 +17,17 @@ use fpdf\Enums\PdfDestination;
 use fpdf\Enums\PdfFontName;
 use fpdf\Enums\PdfFontStyle;
 use fpdf\PdfDocument;
-use fpdf\PdfException;
-use PHPUnit\Framework\Attributes\Depends;
 use PHPUnit\Framework\TestCase;
 
 abstract class AbstractCompareTestCase extends TestCase
 {
-    protected bool $unlink = true;
-
-    private string $newFile = '';
-    private string $oldFile = '';
+    /**
+     * Pattern to remove the creation date and the producer.
+     */
+    private const PATTERN = [
+        '/\/CreationDate.*\)/mi',
+        '/\/Producer \(FPDF.*\)/mi',
+    ];
 
     #[\Override]
     public static function setUpBeforeClass(): void
@@ -36,67 +37,28 @@ abstract class AbstractCompareTestCase extends TestCase
         }
     }
 
-    #[\Override]
-    protected function setUp(): void
-    {
-        $this->newFile = __DIR__ . '/doc_new.pdf';
-        $this->oldFile = __DIR__ . '/doc_old.pdf';
-    }
-
-    #[Depends('testOld')]
-    #[Depends('testNew')]
     public function testEqual(): void
     {
-        self::assertFileExists($this->oldFile);
-        self::assertFileExists($this->newFile);
+        $oldDocument = $this->createOldDocument();
+        $this->updateOldDocument($oldDocument);
+        /** @phpstan-var string $oldContent */
+        $oldContent = $oldDocument->Output('S');
+        $oldContent = \preg_replace(self::PATTERN, '', $oldContent);
+        self::assertIsString($oldContent);
 
-        $old_content = \file_get_contents($this->oldFile);
-        $new_content = \file_get_contents($this->newFile);
+        $newDocument = $this->createNewDocument();
+        $this->updateNewDocument($newDocument);
+        $newContent = $newDocument->output(PdfDestination::STRING);
+        $newContent = \preg_replace(self::PATTERN, '', $newContent);
+        self::assertIsString($newContent);
 
-        if ($this->unlink) {
-            \unlink($this->oldFile);
-            \unlink($this->newFile);
-        }
-
-        self::assertIsString($old_content);
-        self::assertIsString($new_content);
-
-        // remove creation date and producer
-        $patterns = [
-            '/\/CreationDate.*\)/mi',
-            '/\/Producer \(FPDF.*\)/mi',
-        ];
-        $old_content = \preg_replace($patterns, '', $old_content);
-        $new_content = \preg_replace($patterns, '', $new_content);
-
-        self::assertSame($old_content, $new_content);
+        self::assertSame($oldContent, $newContent);
     }
 
-    /**
-     * @throws PdfException
-     */
-    public function testNew(): void
-    {
-        $doc = $this->createNewDocument();
-        $this->updateNewDocument($doc);
-        $doc->output(PdfDestination::FILE, $this->newFile);
-        self::assertFileExists($this->newFile);
-    }
-
-    public function testOld(): void
-    {
-        $doc = $this->createOldDocument();
-        $this->updateOldDocument($doc);
-        $doc->Output('F', $this->oldFile);
-        self::assertFileExists($this->oldFile);
-    }
-
-    /**
-     * @throws PdfException
-     */
     protected function createNewDocument(): PdfDocument
     {
         $doc = new PdfDocument();
+        $doc->setCompression(false);
         $doc->setFont(PdfFontName::ARIAL, PdfFontStyle::REGULAR, 9.0);
         $doc->addPage();
 
@@ -106,15 +68,13 @@ abstract class AbstractCompareTestCase extends TestCase
     protected function createOldDocument(): FPDF
     {
         $doc = new FPDF();
+        $doc->SetCompression(false);
         $doc->SetFont('Arial', '', 9.0);
         $doc->AddPage();
 
         return $doc;
     }
 
-    /**
-     * @throws PdfException
-     */
     abstract protected function updateNewDocument(PdfDocument $doc): void;
 
     abstract protected function updateOldDocument(FPDF $doc): void;